+++ pytest.ini
@@ -1,5 +1,10 @@
 [pytest]
 markers =
-    scenario_test_task: my scenario from test task
+    test_scenario_task: my scenario from test task
 
 python_classes = *Test Test* *Suite
+;testpaths = TestPage
+;addopts = -rsxX -l --tb=short --strict
+;miniversion = 3.0
+;norecursedirs = .* venv src *. dist build
+;xfail_strict = true
